Cleaned up teacher activity
