+++ app/src/main/java/com/example/nathaniel/heyteacher/MainActivity.java
@@ -22,7 +22,7 @@ public class MainActivity extends AppCompatActivity {
 
     public void createRoom(View view){
         //Toast.makeText(this, "Create room", Toast.LENGTH_SHORT).show();
-        Intent intent = new Intent(this, TeacherScreen.class);
+        Intent intent = new Intent(this, TeacherActivity.class);
         intent.putExtra("name", nameField.getText());
         startActivity(intent);
     }
