+++ app/src/main/java/com/example/nathaniel/heyteacher/TeacherActivity.java
@@ -1,0 +1,13 @@
+package com.example.nathaniel.heyteacher;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+public class TeacherActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_teacher);
+    }
+}
